reverted merging two tests
